requirements.txt
numpy>=1.21.0
pyrealsense2>=2.54.0
numba>=0.56.0
cupy-cuda12x>=13.6.0  # use cupy-cuda11x for CUDA 11.x
matplotlib>=3.5.0
open3d>=0.17.0
scipy>=1.9.0
line_profiler>=4.0.0
memory_profiler>=0.60.0
tqdm>=4.64.0
opencv-python>=4.8.0

